// src/app/Interfaces.ts
export interface Person {
    id: string;
    first_name: string;
    last_name: string;
    gender: string;
    age: number;
    major: string;
    occupation: string;
    location: string;
}
export interface Filter {
    firstname: string;
    lastname: string;
    gender: string;
    age: number;
    major: string;
    occupation: string;
    location: string;
}